Refactored view all case notes. Moved the presenter and view classes into a 'viewAll' subfolder.
