+++ server/routes/caseNotes/caseNotesController.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express'
 import InterventionsService from '../../services/interventionsService'
-import CaseNotesPresenter from './caseNotesPresenter'
-import CaseNotesView from './caseNotesView'
+import CaseNotesPresenter from './viewAll/caseNotesPresenter'
+import CaseNotesView from './viewAll/caseNotesView'
 import ControllerUtils from '../../utils/controllerUtils'
 import CommunityApiService from '../../services/communityApiService'
 import HmppsAuthService from '../../services/hmppsAuthService'
